fixed problems with semi-transparent images
fixed situations where semi-transparent images (i.e., if you used a feathered crop in photoshop) would result in large boundaries on the edges of the pattern. Now, anything with an alpha value less than 150 is considered transparent. and ignored. Everything else is turned opaque. Additionally, changed the background that transparent images are merged with to black. This only matters with semi-transparent images, but I should implement being able to choose the background color later.