Now matches DMC in LAB colorspace
Changed the default behavior of the program to match colors in the image to DMC thread colors in LAB colorspace. Because this is more perceptually uniform, the patterns come out looking more like the input image. I used the CMC l:c standard color difference method in scikit-image because it seemed to produce better results than any of the CIEDE distance functions; see the readme for more details!